pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "healthcare-ai-triage"
version = "0.1.0"
description = "Interpretable triage and summarization baseline for the 2026 Healthcare AI hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "huggingface_hub>=0.30.0",
    "pandas>=2.3.0",
    "numpy>=2.3.0",
    "openai>=1.68.0",
    "scikit-learn>=1.7.0",
    "streamlit>=1.43.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
